move statusBar services and extension point from /browser to /common
