+++ src/vs/workbench/contrib/offline/browser/offline.contribution.ts
@@ -14,7 +14,7 @@ import { localize } from 'vs/nls';
 import { combinedDisposable, DisposableStore, IDisposable } from 'vs/base/common/lifecycle';
 import { Event } from 'vs/base/common/event';
 import { DomEmitter } from 'vs/base/browser/event';
-import { IStatusbarService, StatusbarAlignment } from 'vs/workbench/services/statusbar/browser/statusbar';
+import { IStatusbarService, StatusbarAlignment } from 'vs/workbench/services/statusbar/common/statusbar';
 
 export const STATUS_BAR_OFFLINE_BACKGROUND = registerColor('statusBar.offlineBackground', {
 	dark: '#6c1717',
