+++ src/vs/workbench/services/statusbar/common/statusbar.ts
